added classes for svm
